third_party/blink/web_tests/fast/lists/ol-nested-items-dynamic-remove-expected.txt

This tests that list item markers are updated when removing existing items that are not direct children of the list element.

1 One
2 Two
3 Three
4 Four
5 Five
6 Six
7 Seven
